Add a logo
The logo is from
https://commons.wikimedia.org/wiki/File:Water_Python_(Liasis_mackloti)_(8692394648).jpg.
It claims to be Liasis mackloti, whereas the Brown Water Python is Liasis
fuscus. However, from what I can tell, the two are quite similar, and even
http://reptile-database.reptarium.cz/species?genus=Liasis&species=fuscus
claims they might be synonymous. At any rate, I couldn't find a freely
licensed image of the Liasis fuscus, so this will have to do.